task/3- Segregated selenium automation
-- Segregated `SlackUrlUpdater` to
    `automation_model` package's
    `automation` sub-package's  `automate_slack_event_subscription.py` module
